// frontend/src/api/models/handlerInventoryItemResponse.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

/**
 * Inventory item response with stock quantities and cost information
 */
export interface HandlerInventoryItemResponse {
  available_quantity?: number
  created_at?: string
  id?: string
  is_above_maximum?: boolean
  is_below_minimum?: boolean
  locked_quantity?: number
  max_quantity?: number
  min_quantity?: number
  product_id?: string
  tenant_id?: string
  total_quantity?: number
  total_value?: number
  unit_cost?: number
  updated_at?: string
  version?: number
  warehouse_id?: string
}