// frontend/src/api/models/listInventoryBelowMinimumParams.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */
import type { ListInventoryBelowMinimumOrderDir } from './listInventoryBelowMinimumOrderDir'

export type ListInventoryBelowMinimumParams = {
  /**
   * Filter by warehouse ID
   */
  warehouse_id?: string
  /**
   * Page number
   */
  page?: number
  /**
   * Page size
   * @maximum 100
   */
  page_size?: number
  /**
   * Order by field
   */
  order_by?: string
  /**
   * Order direction
   */
  order_dir?: ListInventoryBelowMinimumOrderDir
}