// frontend/src/api/models/listInventoryBelowMinimumOrderDir.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type ListInventoryBelowMinimumOrderDir =
  (typeof ListInventoryBelowMinimumOrderDir)[keyof typeof ListInventoryBelowMinimumOrderDir]

export const ListInventoryBelowMinimumOrderDir = {
  asc: 'asc',
  desc: 'desc',
} as const